+++ HibernateTry/src/main/java/by/moiseenko/service/impl/AuthorServiceImpl.java
@@ -37,7 +37,7 @@ public class AuthorServiceImpl implements AuthorService {
   }
 
   @Override
-  public Author getAuthorById(int id) {
+  public Author getAuthorById(Long id) {
     return authorDAO.findAuthorById(id);
   }
